Change identation

+++ Assets/Scripts/Data/Defender/DefenderData.cs
@@ -1,8 +1,7 @@
 ﻿using UnityEngine;
 
 [CreateAssetMenu(fileName = "Zombie Data", menuName = "Necromancer/Defender/Data", order = 1)]
-public class DefenderData : ScriptableObject
-{
+public class DefenderData : ScriptableObject {
     [Tooltip("The level of the defender")]
     public int Level; // The level of the defender
     [Tooltip("The damage inflicted by the defender")]
@@ -12,12 +11,11 @@ public class DefenderData : ScriptableObject
     [Tooltip("The movement speed of the defender")]
     public int Speed; // The movement speed of the defender
     [Tooltip("The attack speed of the defender")]
-    public int AttackSpeed; // The attack speed of the defender
+    public float AttackSpeed; // The attack speed of the defender
     [Tooltip("The range from which to start attacking")]
     public float AttackRange = 0.5f; // The range from which to start attacking
 
-    public DefenderData()
-    {
+    public DefenderData() {
         Level = 1;
         Damage = 1;
         Health = 1;
